pyproject.toml
[project]
name = "foundry-grant-eo-validation-demo"
version = "0.1.0"
description = "Grant proposal compliance automation using Azure AI Foundry"
authors = [
    {name = "Your Organization", email = "your-email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "streamlit>=1.51.0",
    "python-dotenv>=1.2.1",
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.25.1",
    "azure-ai-inference>=1.0.0b1",
    "azure-search-documents>=11.4.0",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.36.0",
    "azure-storage-blob>=12.19.0",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "requests>=2.32.5",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.1.0",
    "plotly>=6.4.0",
    "fastapi>=0.121.2",
    "agent-framework>=1.0.0b251114",
    "azure-functions-durable>=1.4.0",
    "azure-ai-formrecognizer>=3.3.3",
    "msgraph-sdk>=1.48.0",
    "nest-asyncio>=1.6.0",
    "azure-storage-queue>=12.14.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.14.5",
    "pyright>=1.1.407",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.10"
